reorganized and added a bunch of commands
